// UI-EasyCount/src/components/atoms/Inputs/Input.types.tsx
export enum InputType {
  MealName = "Meal Name",

  Name = "Goal Name",
  Kcal = "Calories Goal",
  Carbs = "Carbohydrates Goal",
  Fat = "Fat Goal",
  Protein = "Protein Goal",

  Search = "Search by name",

  NewFoodName = "New Food Name",
  FoodPortion = "Food Portion",
  KcalPerPortion = "Kcal per Portion",
  CarbsPerPortion = "Carbs per Portion",
  FatPerPortion = "Fat per Portion",
  ProteinPerPortion = "Protein per Portion",
}